requirements.txt
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
fonttools==4.39.3
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pdfminer.six==20221105
Pillow==9.5.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
zipp==3.15.0
